move token description to its own component, add copy, make simple project query
